+++ include/atlas/atlas.h
@@ -32,7 +32,9 @@ class LunarRayTracer {
 
         // Image Generation Routines
         bool generateImageOptical(const std::string& filename, int type = CV_8UC1); 
-        bool generateImageDEM(const std::string& filename, int type = CV_8UC1); 
+        bool generateImageDEM(
+            const std::string& filename, int type = CV_8UC1, bool normalize = true
+        ); 
         bool generateDepthMap(const std::string& filename, int type = CV_8UC1);
 
         void exportRayTracedInfo(const std::string& filename);